// types/global.d.ts
export type Comment = {
  id: string;
  comment: string;
  createdAt: Date;
  // updatedAt: Date;
  feedsPostId: string;
  userId: string;
  replies: Comment[];
  userData: {
    firstName: string;
    lastName: string;
    userName: string;
    profileImage?: string;
  };
};

export type Reply = {
  id: string;
  reply: string;
  feedsPostCommentId: string;
  // feedsPostCommentReplyId: string;
  userId: string;
  createdAt: Date;
  userData: {
    firstName: string;
    lastName: string;
    userName: string;
    profileImage?: string;
  };

};